Updated GUI
Made the Simulink model's runtime into a function of user input, tpulse.

Also got rid of the template's "English vs metric" option. The GUI code
still has "metricdata" sprinkled throughout the code, but I'm not sure
how to get rid of it without breaking the reset button. It works if you
keep it though, so meh.